WIP of executing functions.
I don't think this is going to work though,
because some will need to "run" before they
resolve the rest of their args.

e.g. (let 'a 1 (print a))
Needs to "run" to add "a" to the local scope
it uses to run print.

Also, returning a copy of the new global scope
is probably the better way to go.

Maybe the first step should just be building
call objects and validating the argument types