Fix multiple let values that are Calls picking up new values
.e.g
(let 'a 1
  (let 'a (+ a 1) 'b (+ a 1)
    (print b)
  )
)

This should have:
a=1
a=a+1=2 b=a+1=2
print b

However, we used the updated local scope for the second assignment.
Meaning that the second call for 'b (+ a 1) ended up as 2+1, not 1+1.

To fix this take a copy of the scope before we add anything to the
new scope. Use that copy to evaluate any calls.

Side note:
There isn't an issue for simply referencing a symbol since that is
done in exec inner before the let does anything. Like:
(let 'a 1
  (let 'a (+ a 1) 'b a
    (print b)
  )
)

b=a is resolved before we enter let's breadth processing function.