lambdas part 2
Got argument validation written but untested.

This is because of the other builtin executors
being function pointers. Whereas the user defined
executor is a closure.

If the closure doesn't reference the current scope,
it compiles. (decays to a fn pointer?)
If it captures (Function itself and local_scope here)
it won't, meaning it cannot be stored as a function pointer.