Add specific "declared but not defined" errors
These will always result from letrec.

search_scope now returns a double Option.
Where the first is whether the name existed
and the second is the Option<ASTType>.

That's how we get the different errors.
If the first Option is None then it's undeclared/unknown symbol.
If the second Option is None then it's declared but undefined symbol.
If both are Some then you found your symbol.