Add call stacks to most errors on ASTTypes
Most because the comparisons won't have it.

For these errors we print the callstack
then print the location again for the error.
Since the last call and the error location
won't be the same.

Doing this is super tedious. The real soloution
is to return result types from, well, everything.
Then we can choose to panic later at a point
where we have all the info.

Instead of passing all this state around.