Include builtins in symbol lookup
Meaning you can do:
(map + (list 1 2 3)

This is done by wrapping the name in a new type,
BuiltinFunctionWrapper. This name will be prepended
with "__builtin_".

Then since this wrapper will be found in local scope
when we try to call it, we will treat it as user defined
and end up in builtin_user_defined_function.

Here, if we've got a wrapper, instead of calling the
user's function, we'll make on the fly a call object and
exec that. So if you have:
(let 'fn + ...
There'll be a BuiltinFunctionWrapper for "__builtin_+".
Then later:
(fn 1 2)
We get to builtin_user_defined_function and see we got a
BuiltinFunctionWrapper. So we build a call that looks like:
(__builtin_+ 1 2)
Then exec it.

Here's where the prefix comes in. If something shadowed +
between the time we built the wrapper and we called it,
the + in the ususal scope lookup will be that new +.

So with the ___builtin_ prefix we know it will never be
found in the local or global scope and I've changed the
builtin lookup to remove this prefix if we find it.

That way we always get the builtin + no matter what is going
on.

All of which means "__" prefix symbols are now reserved.
I always planned to do this anyway.