Reduce number of stack frames used for split
By writing it in terms of split once while we
have characters left to split.

Meaning after each split point is found we can
"return" some frames instead of going deeper
each time.

So we're still piling up frames but this time
it scales with the number of characters between
each split point. Rather than with the length of
the entire string.

This means we can take the align demo input as
one long string and split on newlines.

Again, this doesn't actually fix the recursion
limits but it's good enough for now.